backend/requirements.txt
fastapi
uvicorn
google-generativeai
google-genai
openai
python-dotenv
pydantic
